+++ docx/oxml/table.py
@@ -771,7 +771,14 @@ class CT_TcPr(BaseOxmlElement):
     gridSpan = ZeroOrOne('w:gridSpan', successors=_tag_seq[3:])
     vMerge = ZeroOrOne('w:vMerge', successors=_tag_seq[5:])
     vAlign = ZeroOrOne('w:vAlign', successors=_tag_seq[12:])
+    shd = ZeroOrOne('w:shd')
     del _tag_seq
+
+    @property
+    def background_color(self):
+        if self.shd is not None:
+            return self.shd.fill
+        return None
 
     @property
     def grid_span(self):