+++ docx/picture.py
@@ -17,6 +17,16 @@ class PictureShape(Parented):
     def rId(self):
         return self._pic.blipFill.blip.embed
 
+    @property
+    def has_wrap(self):
+        if hasattr(self.parent,'has_wrap'):
+            return self.parent.has_wrap
+        return None
+
+    @property
+    def parent(self):
+        return self._pic.parent
+
     @lazyproperty
     def parent_section(self):
         parent = self._pic.getparent()
@@ -34,14 +44,14 @@ class PictureShape(Parented):
 
     @lazyproperty
     def off_x(self):
-        off_x = self._pic.spPr.ox.pt
+        off_x = self._pic.spPr.ox.pt + self._pic.parent_off_x.pt
         if isinstance(self.parent_section, TextboxContent):
             off_x += self.parent_section.off_x
         return off_x
 
     @lazyproperty
     def off_y(self):
-        off_y = self._pic.spPr.ox.pt
+        off_y = self._pic.spPr.ox.pt + self._pic.parent_off_y.pt
         if isinstance(self.parent_section, TextboxContent):
             off_y += self.parent_section.off_y
         return off_y